creates solutions folder in components
