+++ src/app/solutions/belus.tsx
@@ -13,11 +13,11 @@ import Animated, {
 } from 'react-native-reanimated'
 import { Gesture, GestureDetector } from 'react-native-gesture-handler'
 import { Colors } from '../../theme/colors'
-import { Slimy } from '../../components/BelusChallenge/Slimy'
-import { styles } from '../../components/BelusChallenge/styles'
+import { Slimy } from '../../components/solutions/BelusChallenge/Slimy'
+import { styles } from '../../components/solutions/BelusChallenge/styles'
 import { clamp, mix } from 'react-native-redash'
-import { useGeneralDimensions } from '../../hooks/useGeneralDimensions'
-import { Lives } from '../../components/BelusChallenge/Lives'
+import { useGeneralDimensions } from '../../components/solutions/hooks/useGeneralDimensions'
+import { Lives } from '../../components/solutions/BelusChallenge/Lives'
 import { Stack } from 'expo-router'
 import Avocado from '../../../assets/images/food/avocado.png'
