+++ src/app/solutions/fourth.tsx
@@ -11,10 +11,10 @@ import {
 } from 'react-native-reanimated'
 import { Gesture, GestureDetector } from 'react-native-gesture-handler'
 import { Colors } from '../../theme/colors'
-import { Slimy } from '../../components/FourthChallenge/Slimy'
-import { styles } from '../../components/FourthChallenge/styles'
+import { Slimy } from '../../components/solutions/FourthChallenge/Slimy'
+import { styles } from '../../components/solutions/FourthChallenge/styles'
 import { clamp } from 'react-native-redash'
-import { useGeneralDimensions } from '../../hooks/useGeneralDimensions'
+import { useGeneralDimensions } from '../../components/solutions/hooks/useGeneralDimensions'
 import { Stack } from 'expo-router'
 
 interface FourthChallengeProps {
