+++ src/components/solutions/BelusChallenge/Lives/index.tsx
