+++ src/components/solutions/BelusChallenge/Slimy/Mouth/index.tsx
