+++ src/components/solutions/BelusChallenge/Slimy/XL/index.tsx
