+++ src/components/solutions/BelusChallenge/styles.ts
