+++ src/components/solutions/FifthChallenge/Slimy/XL/index.tsx
