+++ src/components/solutions/FifthChallenge/styles.ts
