+++ src/components/solutions/FourthChallenge/Slimy/XL/index.tsx
