+++ src/components/solutions/FourthChallenge/styles.ts
