+++ src/components/solutions/SecondChallenge/styles.ts
