+++ src/components/solutions/ThirdChallenge/styles.ts
