+++ src/components/solutions/hooks/useGeneralDimensions.ts
